src/auth_services/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.10
mysql-connector-python==8.1.0
PyJWT==2.3.0
python-dotenv==1.0.0